requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cupy-cuda12x==13.6.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fastrlock==0.8.3
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
iniconfig==2.3.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.3
packaging==25.0
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.1.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.4
pytest==9.0.2
python-dateutil==2.9.0.post0
pywin32==311
pyzmq==27.1.0
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
torch==2.8.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.2.13
